refactor(server): fold hooks.ts into types.ts
`app/hooks.ts` held four derived hook-context types
(`DispatchAuthorizeContext`, `DispatchAdmissionResult`,
`MessageAuthorizeContext`, `MessageAuthorizeResult`) consumed only by
`app-host.ts` and its two test files. They are server-side hook types
that belong next to the existing `ConnectionHook`/`DisconnectionHook`
types in `types.ts`. Fold them in and repoint the three importers; the
`ParamsOf<DispatchAuthorize/MessagesAuthorize>` derivation is unchanged.
