+++ packages/server/src/app/app-host.remote.test.ts
@@ -41,7 +41,7 @@ import { AppHost } from "./app-host.js";
 import type {
   MessageAuthorizeContext,
   DispatchAuthorizeContext,
-} from "./hooks.js";
+} from "./types.js";
 
 const liveIt = effectIt.live;
